Adding some search methods for to child for child node names and attributes
The methods has_child_node? and child_nodes_have_attribute? will traverse the
xml document starting from the node referenced looking for a node with the name
given or a node with the attribute given.  It will not tell you where the node
or attribute is located, it will just give you the boolean indicating that they
do exist somewhere in the document.